% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_correlation.R
\name{correlationPlot}
\alias{correlationPlot}
\title{A correlation plot}
\usage{
correlationPlot(
  plotData,
  xVar,
  yVar,
  xVarTitle = "Estimated Theta",
  yVarTitle = "True Theta",
  xRange = c(floor(min(c(plotData[[xVar]], plotData[[yVar]]))),
    ceiling(max(c(plotData[[xVar]], plotData[[yVar]])))),
  opacity = 0.4,
  addDiagLine = F,
  useDiffAsColor = T,
  cutsAt = NULL,
  addTitle = T,
  plotName = ""
)
}
\description{
A correlation plot
}
