% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_formAnalysis.R
\name{formAnalysis_mod_server}
\alias{formAnalysis_mod_server}
\title{formAnalysis module server-side processing}
\usage{
formAnalysis_mod_server(input, output, session, items, forms, colMap)
}
\arguments{
\item{input, output, session}{are standard parameters necessary for Shiny modules.}

\item{items}{the item pool}

\item{forms}{the forms}

\item{colMap}{the col
 a list of IRT related column information. An example:
IRTDataConf = list(IRTParCols = c("IRT_PAR_A", "IRT_PAR_B", "IRT_PAR_C"), IRTModelCol = "IRT_MODEL",
IRTScale = 1.0, theta = c(-4,4))}
}
\description{
formAnalysis module server-side processing
}
