% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_formAnalysis.R
\name{formAnalysis_mod_ui}
\alias{formAnalysis_mod_ui}
\title{formAnalysis module user interface}
\usage{
formAnalysis_mod_ui(id)
}
\arguments{
\item{id}{a namespace id}
}
\value{
a tagList output
}
\description{
formAnalysis module user interface
}
