% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_IRT_plot.R
\name{gen_formIRTPlot}
\alias{gen_formIRTPlot}
\title{Generate IRT plots given forms}
\usage{
gen_formIRTPlot(
  metadata,
  formInf,
  joinBy,
  IRT_fun = "Information",
  theta = c(-4, 4),
  IRTParCols,
  IRTModelCol,
  IRTScale = 1,
  showLegend = T,
  opacity = 0.6,
  plotType = c("curve", "")
)
}
\arguments{
\item{metadata}{a data frame contains IRT parameters and categorical attributes}

\item{formInf}{form information in tibble that must contain columns named form_id and item_id}

\item{joinBy}{join for formInf and metadata}

\item{IRT_fun}{one or multiple IRT plot names, which could be "Information", "CSEM",
"Expected_Percent_Score", or/and "Expected_Raw_Score".}

\item{theta}{A vector of ability values}

\item{IRTParCols}{A vector of colnames that have the columns required by the IRT model of choice.}

\item{IRTModelCol}{A column that specifies the IRT model for each item.
This can also be a single value, which will be recycled.}

\item{IRTScale}{A vector of scaling parameters used in the IRT models.
This can also be a single value, which will be recycled.}
}
\description{
The supported IRT plots include "Information", "CSEM", "Expected_Percent_Score",
"Expected_Raw_Score".
}
