% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_IRT_plot.R
\name{gen_IRT_plot}
\alias{gen_IRT_plot}
\title{Generate IRT plot(s)}
\usage{
gen_IRT_plot(
  data,
  IRT_fun = "Information",
  theta = c(-4, 4),
  IRTParCols,
  IRTModelCol,
  IRTScale = 1,
  groupByCol = NULL,
  colorByCol = NULL,
  showLegend = T,
  opacity = 0.6
)
}
\arguments{
\item{data}{a data frame contains IRT parameters and categorical attributes.}

\item{IRT_fun}{one or multiple IRT plot names, which could be "Information", "CSEM",
"Expected_Percent_Score", or/and "Expected_Raw_Score".}

\item{theta}{A vector of ability values}

\item{IRTParCols}{A vector of colnames that have the columns required by the IRT model of choice.}

\item{IRTModelCol}{A column that specifies the IRT model for each item.
This can also be a single value, which will be recycled.}

\item{IRTScale}{A vector of scaling parameters used in the IRT models.
This can also be a single value, which will be recycled.}

\item{groupByCol}{a column that groups the data for IRT functions and assigns plot colors}

\item{colorByCol}{a column that uses for colors. To use colorByCol, the groupByCol must not be
NULL and the number of unique values in groupByCol must be more than the number of unique values
in colorByCol.}
}
\description{
The supported IRT plots include "Information", "CSEM", "Expected_Percent_Score",
"Expected_Raw_Score".
}
