% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_IRTPlot.R
\name{IRTPlot_mod_ui}
\alias{IRTPlot_mod_ui}
\title{IRTPlot module user interface}
\usage{
IRTPlot_mod_ui(id)
}
\arguments{
\item{id}{a namespace id}
}
\value{
a plot_ly output
}
\description{
IRTPlot module user interface
}
