% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_isr_plot.R
\name{isr_plot_server}
\alias{isr_plot_server}
\title{ISR plot module server side function}
\usage{
isr_plot_server(input, output, session, dataset, plot1vars, plot2vars)
}
\arguments{
\item{input, output, session}{standard \code{shiny} boilerplate}

\item{dataset}{data frame (non-reactive) with variables necessary for scatterplot}

\item{plot1_vars}{list containing reactive x-variable name (called `xvar`) and y-variable name (called `yvar`) for plot 1}

\item{plot2_vars}{list containing reactive x-variable name (called `xvar`) and y-variable name (called `yvar`) for plot 2}
}
\description{
This module produces a scatterplot with the sales price against a variable selected by the user.
}
