% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_isr_plot.R
\name{isr_plot_ui}
\alias{isr_plot_ui}
\title{ISR plot module UI function}
\usage{
isr_plot_ui(id)
}
\arguments{
\item{id, }{character used to specify namespace, see \code{shiny::\link[shiny]{NS}}}
}
\description{
ISR plot module UI function
}
