% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runShiny_editDT.R
\name{runShiny_editDT}
\alias{runShiny_editDT}
\title{A shiny app for testing the editDT module module}
\usage{
runShiny_editDT()
}
\value{
a shiny app "runShiny_editDT"
}
\description{
A shiny app for testing the editDT module module
}
\examples{
\dontrun{ runShiny_editDT() }

}
