% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runShiny_editDT_V2.R
\name{runShiny_editDT2}
\alias{runShiny_editDT2}
\title{A shiny app for testing the editDT2 module module}
\usage{
runShiny_editDT2()
}
\value{
a shiny app "runShiny_editDT2"
}
\description{
A shiny app for testing the editDT2 module module
}
\examples{
\dontrun{ runShiny_editDT2() }

}
