% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runShiny_formAnalysis.R
\name{runShiny_formAnalysis}
\alias{runShiny_formAnalysis}
\title{A shiny app for testing IRT plot module}
\usage{
runShiny_formAnalysis()
}
\value{
a shiny app "runShiny_formAnalysis"
}
\description{
A shiny app for testing IRT plot module
}
\examples{
\dontrun{ runShiny_formAnalysis() }
}
