% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runShiny_IRTPlot.R
\name{runShiny_IRTPlot}
\alias{runShiny_IRTPlot}
\title{A shiny app for testing IRT plot module}
\usage{
runShiny_IRTPlot()
}
\value{
a shiny app "runShiny_IRTPlot"
}
\description{
A shiny app for testing IRT plot module
}
\examples{
\dontrun{ runShiny_IRTPlot() }
}
