% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runShiny_isr_plot.R
\name{runShiny_isr_plot}
\alias{runShiny_isr_plot}
\title{A shiny app for testing ISR plot module}
\usage{
runShiny_isr_plot()
}
\value{
a shiny app "runShiny_isr_plot"
}
\description{
A shiny app for testing ISR plot module
}
\examples{
\dontrun{ runShiny_isr_plot() }
}
